FIX: Replace RAND_bytes with CryptographicRandom::fillBuffer
https://www.openssl.org/docs/man1.1.1/man3/RAND_bytes.html:
If the entropy source fails or is not available, the CSPRNG will enter an error state and refuse to generate random bytes. For that reason, it is important to always check the error return value of RAND_bytes() and RAND_priv_bytes() and not take randomness for granted.

Also remove unnecessary cast.

+++ src/crypto/CryptographicRandom.cpp
@@ -24,7 +24,7 @@ void CryptographicRandom::fillBuffer(void *buf, int numBytes) {
 	// OpenSSL needs at least 32-bytes of high-entropy random data to seed its CSPRNG.
 	// If OpenSSL cannot acquire enough random data to seed its CSPRNG at the time Mumble and Murmur
 	// are running, there is not much we can do about it other than aborting the program.
-	if (RAND_bytes(reinterpret_cast< unsigned char * >(buf), static_cast< int >(numBytes)) != 1) {
+	if (RAND_bytes(reinterpret_cast< unsigned char * >(buf), numBytes) != 1) {
 		qFatal("CryptographicRandom::fillBuffer(): internal error in OpenSSL's RAND_bytes or entropy pool not yet "
 			   "filled.");
 	}
